BudgetServiceTest approved

+++ src/test/java/com/solvd/it/company/CostEstimatorTest.java
@@ -22,9 +22,9 @@ class CostEstimatorTest {
     @Test
     @DisplayName("estimateCost_validInputs_returnsExpectedCost")
     void estimateCost_validInputs_returnsExpectedCost() throws InvalidHoursPerDayException {
-        when(mockTeam.getAverageHourlyRate()).thenReturn(50.0f);
-        when(mockTeam.getTeamSize()).thenReturn(5);
-        when(mockProjectProcess.getHoursPerWeek()).thenReturn(4);
+        when(mockTeam.getAverageHourlyRate()).thenReturn(Float.valueOf(50.0f));
+        when(mockTeam.getTeamSize()).thenReturn(Integer.valueOf(5));
+        when(mockProjectProcess.getHoursPerWeek()).thenReturn(Integer.valueOf(4));
 
         float result = CostEstimator.estimateCost(mockTeam, mockProjectProcess, 8);
 
@@ -53,9 +53,9 @@ class CostEstimatorTest {
     @Test
     @DisplayName("estimateCost_zeroHoursPerDay_returnsZeroCost")
     void estimateCost_zeroHoursPerDay_returnsZeroCost() throws InvalidHoursPerDayException {
-        when(mockTeam.getAverageHourlyRate()).thenReturn(100.0f);
-        when(mockTeam.getTeamSize()).thenReturn(3);
-        when(mockProjectProcess.getHoursPerWeek()).thenReturn(5);
+        when(mockTeam.getAverageHourlyRate()).thenReturn(Float.valueOf(100.0f));
+        when(mockTeam.getTeamSize()).thenReturn(Integer.valueOf(3));
+        when(mockProjectProcess.getHoursPerWeek()).thenReturn(Integer.valueOf(5));
 
         float result = CostEstimator.estimateCost(mockTeam, mockProjectProcess, 0);
         assertEquals(0.0f, result);
@@ -64,11 +64,13 @@ class CostEstimatorTest {
     @Test
     @DisplayName("estimateCost_maxValidHours_returnsHighCost")
     void estimateCost_maxValidHours_returnsHighCost() throws InvalidHoursPerDayException {
-        when(mockTeam.getAverageHourlyRate()).thenReturn(75.0f);
-        when(mockTeam.getTeamSize()).thenReturn(10);
-        when(mockProjectProcess.getHoursPerWeek()).thenReturn(5);
+        when(mockTeam.getAverageHourlyRate()).thenReturn(Float.valueOf(75.0f));
+        when(mockTeam.getTeamSize()).thenReturn(Integer.valueOf(10));
+        when(mockProjectProcess.getHoursPerWeek()).thenReturn(Integer.valueOf(5));
 
         float result = CostEstimator.estimateCost(mockTeam, mockProjectProcess, 24);
         assertEquals(90000.0f, result);
     }
 }
+
+// Co-generated with AI assistance (GPT-5), reviewed and validated by Alejandro Ramirez
